Contact form redesign. Added redux-forms.
