+++ package.json
@@ -52,6 +52,7 @@
     "react-slick": "^0.14.5",
     "react-thunk": "^1.0.0",
     "redux": "^3.0.5",
+    "redux-form": "^6.5.0",
     "redux-react-router": "^1.0.0-beta3",
     "redux-thunk": "^2.1.0",
     "rss": "^1.2.2",
