+++ public/css/style.css
@@ -2107,12 +2107,23 @@ button.slick-arrow.slick-prev {
 *****************************************************************************/
 
 .contact-form {
-    background-color: #eee;
     margin: auto;
     margin-top: 10px;
-    border: 2px solid #888;
     border-radius: 4px;
     max-width: 600px;
+    background: #FAFAFA;
+    padding: 10px;
+    border: 1px solid #D7D9D9;
+}
+
+.contact-form input[type="text"],
+.contact-form input[type="email"],
+.contact-form textarea {
+    padding: 12px 16px;
+    background: #FFFFFF;
+    border: 1px solid #D7D9D9;
+    border-radius: 4px;
+    width: 100%;
 }
 
 .contact-form-top-container {
@@ -2161,17 +2172,27 @@ button.slick-arrow.slick-prev {
 }
 
 .contact-send {
-    margin: 15px 5px 15px 5px;
+    border: 0px;
+    background: #F0D943;
+    border-radius: 4px;
     width: 100%;
-    border-width: 2px;
-    border-radius: 2px;
+    font-size: 16px;
+    color: #333;
+}
+
+.contact-label {
+    font-size: 14px;
+    color: #575959;
+    letter-spacing: 0;
+    margin-top: 2px;
+    margin-bottom: 6px;
 }
 
 .contact-status {
     border-width: 1px;
     height: 70px;
     width: 100%;
-    padding: 25 5 5 5px;
+    padding: 25px 5px 5px 5px;
     text-align: center;
 }
