+++ shared/reducers/index.js
